Fixes for Jetson Nano board support.

+++ ggml/src/ggml-cuda/convert.cu
@@ -663,7 +663,7 @@ to_fp16_cuda_t ggml_get_to_fp16_cuda(ggml_type type) {
         case GGML_TYPE_F32:
             return convert_unary_cont_cuda<float>;
         case GGML_TYPE_BF16:
-            return convert_unary_cont_cuda<nv_bfloat16>;
+            return convert_unary_cont_cuda<half>;
         default:
             return nullptr;
     }
@@ -712,7 +712,7 @@ to_fp32_cuda_t ggml_get_to_fp32_cuda(ggml_type type) {
         case GGML_TYPE_F16:
             return convert_unary_cont_cuda<half>;
         case GGML_TYPE_BF16:
-            return convert_unary_cont_cuda<nv_bfloat16>;
+            return convert_unary_cont_cuda<half>;
         default:
             return nullptr;
     }
@@ -723,7 +723,7 @@ to_fp16_nc_cuda_t ggml_get_to_fp16_nc_cuda(ggml_type type) {
         case GGML_TYPE_F32:
             return convert_unary_cuda<float>;
         case GGML_TYPE_BF16:
-            return convert_unary_cuda<nv_bfloat16>;
+            return convert_unary_cuda<half>;
         default:
             return nullptr;
     }
